+++ SinZsEventTester/Framework/ModConfig.cs
@@ -5,7 +5,12 @@ public sealed class ModConfig
 {
     public int EventSpeedRatio { get; set; } = 4;
 
-    public int FastForwardRatio { get; set; } = 7;
+    private int fastForwardRatio = 7;
+    public int FastForwardRatio
+    {
+        get => this.fastForwardRatio;
+        set => this.fastForwardRatio = Math.Max(1, value);
+    }
 
     public KeybindList FastForwardKeybind { get; set; } = new(SButton.K);